apollo-server-core: move logging into separate file
